packages/skeleton-generator/scripts/complete_skeleton_generator.py
#!/usr/bin/env python3
"""
[CREATE] Complete Skeleton Generator Implementation
Agent: GrokIA
Timestamp: 2025-12-03T15:36:00Z

Complete implementation with all missing methods for the agent skeleton generator.
"""

from __future__ import annotations

import json
import logging
import os
import shutil
from datetime import datetime, timezone
from pathlib import Path
from typing import Any, Dict, List, Optional

import yaml
from dataclasses import dataclass, field
from enum import Enum, auto


class ComponentType(Enum):
    """Types of components in the skeleton."""
    TRAINING = auto()
    RULES = auto()
    METHODS = auto()
    DATABASE = auto()
    MEMORY = auto()
    FILES = auto()


@dataclass
class SkeletonConfig:
    """Configuration for skeleton generation."""
    agent_id: str
    timestamp: str
    template_type: str = "default"
    custom_components: List[str] = field(default_factory=list)
    skip_existing: bool = True
    verbose: bool = True


class CompleteAgentSkeletonGenerator:
    """Complete generator for agent skeleton structures."""

    def __init__(self, base_path: Path = None):
        """Initialize the skeleton generator."""
        self.base_path = base_path or Path("skeleton-generator")
        self.config_path = self.base_path / "configs"
        self.templates_path = self.base_path / "templates"
        self.agents_path = self.base_path / "agents"

        # Load agent templates
        self.agent_templates = self._load_agent_templates()

        # Setup logging
        logging.basicConfig(
            level=logging.INFO if os.getenv("VERBOSE", "false") == "true" else logging.WARNING,
            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
        )
        self.logger = logging.getLogger("skeleton.generator")

    def _load_agent_templates(self) -> Dict[str, Any]:
        """Load agent template configurations."""
        template_file = self.config_path / "agent_templates.yaml"
        if template_file.exists():
            with open(template_file) as f:
                return yaml.safe_load(f)
        return {"defaults": {}}

    def _get_agent_config(self, agent_id: str) -> Dict[str, Any]:
        """Get configuration for a specific agent."""
        templates = self.agent_templates.get("agent_templates", {})
        return templates.get(agent_id, self.agent_templates.get("defaults", {}))

    def _generate_timestamp(self) -> str:
        """Generate timestamp in ISO format."""
        return datetime.now(timezone.utc).strftime("%Y%m%d_%H%M%S")

    def _create_memory_files(self, memory_dir: Path, agent_config: Dict[str, Any]) -> None:
        """Create memory component files."""

        # Create knowledge base directory
        knowledge_dir = memory_dir / "knowledge_base"
        knowledge_dir.mkdir(exist_ok=True)

        # Create embeddings directory
        embeddings_dir = memory_dir / "embeddings"
        embeddings_dir.mkdir(exist_ok=True)

        # Create context directory
        context_dir = memory_dir / "context"
        context_dir.mkdir(exist_ok=True)

        # Create memory configuration
        memory_config_file = memory_dir / "config.yaml"
        memory_config_content = {
            "agent_config": agent_config.get("memory_config", {}),
            "system": {
                "type": "vector_database",
                "backend": "chroma",
                "collection_name": f"{agent_config.get('display_name', 'Agent')}_memory"
            },
            "embedding": {
                "model": "all-MiniLM-L6-v2",
                "dimension": 384,
                "cache_enabled": True
            },
            "retrieval": {
                "top_k": 10,
                "similarity_threshold": 0.7,
                "context_window": agent_config.get("memory_config", {}).get("context_window", 8192)
            },
            "consolidation": {
                "enabled": True,
                "interval_hours": 24,
                "max_memories": 1000
            }
        }

        with open(memory_config_file, "w") as f:
            yaml.dump(memory_config_content, f, default_flow_style=False)

        # Create memory tracker
        memory_tracker_file = context_dir / "memory_tracker.json"
        memory_tracker_content = {
            "total_memories": 0,
            "last_consolidation": datetime.now(timezone.utc).isoformat(),
            "memory_categories": {
                "context": 0,
                "learned_patterns": 0,
                "user_preferences": 0,
                "error_recovery": 0
            },
            "performance_metrics": {
                "retrieval_accuracy": 0.0,
                "context_relevance": 0.0,
                "memory_utilization": 0.0
            }
        }

        with open(memory_tracker_file, "w") as f:
            json.dump(memory_tracker_content, f, indent=2)

    def _create_files_structure(self, files_dir: Path) -> None:
        """Create files component structure."""

        # Create subdirectories for different file types
        subdirs = [
            "generated",
            "processed",
            "outputs",
            "cache",
            "temp",
            "backups"
        ]

        for subdir in subdirs:
            (files_dir / subdir).mkdir(exist_ok=True)

        # Create file organization config
        file_config_file = files_dir / "organization.yaml"
        file_config_content = {
            "directories": {
                "generated": "Files created by the agent",
                "processed": "Files that have been processed/modified",
                "outputs": "Final output files and results",
                "cache": "Cached data and temporary files",
                "temp": "Temporary files (auto-cleaned)",
                "backups": "Backup copies of important files"
            },
            "naming_convention": {
                "pattern": "{agent_id}_{timestamp}_{type}_{counter}.{ext}",
                "example": "grokia_20251203_codegen_001.py"
            },
            "cleanup": {
                "temp_files_ttl_hours": 24,
                "cache_files_ttl_days": 7,
                "backup_retention_days": 30
            }
        }

        with open(file_config_file, "w") as f:
            yaml.dump(file_config_content, f, default_flow_style=False)

    def _create_database_files(self, database_dir: Path, agent_config: Dict[str, Any]) -> None:
        """Create database component files."""

        # Create schemas directory
        schemas_dir = database_dir / "schemas"
        schemas_dir.mkdir(exist_ok=True)

        # Create migrations directory
        migrations_dir = database_dir / "migrations"
        migrations_dir.mkdir(exist_ok=True)

        # Create models directory
        models_dir = database_dir / "models"
        models_dir.mkdir(exist_ok=True)

        # Create database configuration
        db_config_file = database_dir / "config.yaml"
        db_config_content = {
            "database": {
                "type": "sqlite",
                "path": "./data/agent_database.db",
                "backup_path": "./backups/"
            },
            "connection": {
                "pool_size": 10,
                "max_overflow": 20,
                "pool_timeout": 30
            },
            "caching": {
                "enabled": True,
                "ttl": 3600,
                "max_size": 1000
            }
        }

        with open(db_config_file, "w") as f:
            yaml.dump(db_config_content, f, default_flow_style=False)

    def generate_complete_skeleton(self, agent_id: str, timestamp: str = None, **kwargs) -> Path:
        """Generate complete agent skeleton with all components."""

        # Use provided timestamp or generate new one
        if timestamp is None:
            timestamp = self._generate_timestamp()

        # Create configuration
        config = SkeletonConfig(
            agent_id=agent_id,
            timestamp=timestamp,
            **kwargs
        )

        self.logger.info(f"Generating complete skeleton for agent: {agent_id}")

        # Create agent directory
        agent_dir = self.agents_path / agent_id
        agent_dir.mkdir(exist_ok=True)

        # Create timestamp subdirectory
        timestamp_dir = agent_dir / timestamp
        if timestamp_dir.exists() and config.skip_existing:
            self.logger.warning(f"Skipping existing directory: {timestamp_dir}")
            return timestamp_dir

        timestamp_dir.mkdir(exist_ok=True)

        # Create component directories
        components = [
            ComponentType.TRAINING,
            ComponentType.RULES,
            ComponentType.METHODS,
            ComponentType.DATABASE,
            ComponentType.MEMORY,
            ComponentType.FILES
        ]

        for component in components:
            component_dir = timestamp_dir / component.name.lower()
            component_dir.mkdir(exist_ok=True)
            self.logger.debug(f"Created directory: {component_dir}")

        # Get agent configuration
        agent_config = self._get_agent_config(agent_id)

        # Create all component files
        self._create_memory_files(timestamp_dir / "memory", agent_config)
        self._create_files_structure(timestamp_dir / "files")

        self.logger.info(f"Complete skeleton generated successfully: {timestamp_dir}")
        return timestamp_dir


def demo_complete_skeleton():
    """Demonstrate the complete skeleton generator."""
    print("🚀 Complete Skeleton Generator Demo")
    print("=" * 50)

    # Initialize generator
    generator = CompleteAgentSkeletonGenerator()

    # Generate skeleton for different agents
    agents = ["grokia", "claude_code", "gemini_flash"]

    for agent_id in agents:
        print(f"\n📁 Generating skeleton for: {agent_id}")
        try:
            skeleton_path = generator.generate_complete_skeleton(agent_id)
            print(f"   ✅ Created: {skeleton_path}")
        except Exception as e:
            print(f"   ❌ Error: {e}")

    print("\n🎉 Demo completed!")
    print(f"📂 Check generated skeletons in: {generator.agents_path}")


if __name__ == "__main__":
    demo_complete_skeleton()
