packages/skeleton-generator/pyproject.toml
[project]
name = "eudorax-skeleton-generator"
version = "1.0.0"
description = "Agent scaffolding generator for EudoraX"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "eudorax-core",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
